clearing state when group is created
